pyproject.toml
[tool.poetry]
name = "ToDo"
version = "0.1.0"
description = "My ToDo App"
authors = ["Feng"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "1.1.2"
python-dotenv = "0.12.0"
Werkzeug = "1.0.1"
requests = "2.23.0"
python-dateutil = "2.8.1"
gunicorn = "^20.0.4"
oauthlib = "3.1.0"
flask-login = "0.5.0"

[tool.poetry.dev-dependencies]
pytest = "5.2"
webdriver-manager = "3.2.2"
selenium = "3.141.0"
pymongo = "3.11.3"
dnspython = "1.16.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"